notebooks/multi_agent_capstone_framework/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlmodel
requests
google-genai==0.4.0  # google genai SDK (Gemini)
google-cloud-aiplatform>=1.38.0